+++ Cargo.toml
@@ -67,6 +67,7 @@ fnv = "1.0.7"
 winapi = "0.3.9"
 
 [dev-dependencies]
+binggan = "0.5.1"
 rand = "0.8.5"
 maplit = "1.0.2"
 matches = "0.1.9"
@@ -143,3 +144,8 @@ harness = false
 [[bench]]
 name = "index-bench"
 harness = false
+
+[[bench]]
+name = "agg_bench"
+harness = false
+
